// .storyblok/types/337287/storyblok-components.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { StoryblokAsset } from '../storyblok.d.ts';
export interface Config {
  header_menu?: MenuLink[];
  component: "config";
  _uid: string;
  [k: string]: unknown;
}

export interface FullscreenImage {
  image: StoryblokAsset;
  alt: string;
  component: "fullscreen_image";
  _uid: string;
  [k: string]: unknown;
}

export interface GalleryItem {
  image?: StoryblokAsset;
  alt?: string;
  description?: string;
  component: "gallery_item";
  _uid: string;
  [k: string]: unknown;
}

export interface GallerySection {
  items: GalleryItem[];
  component: "gallery_section";
  _uid: string;
  [k: string]: unknown;
}

export interface Hero {
  headline: string;
  subheadline: string;
  image: StoryblokAsset;
  component: "hero";
  _uid: string;
  [k: string]: unknown;
}

export interface HeroSlider {
  heroes?: Hero[];
  component: "hero_slider";
  _uid: string;
  [k: string]: unknown;
}

export interface ImageTextBlock {
  title: string;
  image: StoryblokAsset;
  description?: string;
  link?: string;
  thumbnail?: StoryblokAsset;
  component: "image_text_block";
  _uid: string;
  [k: string]: unknown;
}

export interface ImageTextGrid {
  items: ImageTextBlock[];
  component: "image_text_grid";
  _uid: string;
  [k: string]: unknown;
}

export interface InfoItem {
  title?: string;
  description?: string;
  component: "info_item";
  _uid: string;
  [k: string]: unknown;
}

export interface InfoSection {
  info?: InfoItem[];
  component: "info_section";
  _uid: string;
  [k: string]: unknown;
}

export interface MenuLink {
  slug?: string;
  label?: string;
  sub_menu?: MenuLink[];
  component: "menu_link";
  _uid: string;
  [k: string]: unknown;
}

export interface Page {
  body?: (
    | Config
    | FullscreenImage
    | GalleryItem
    | GallerySection
    | Hero
    | HeroSlider
    | ImageTextBlock
    | ImageTextGrid
    | InfoItem
    | InfoSection
    | MenuLink
    | Page
  )[];
  component: "page";
  _uid: string;
  [k: string]: unknown;
}